[docker] refs #1433 Change integration-test-disable-wallet-api

+++ ci-scripts/integration-test-disable-wallet-api.sh
@@ -6,16 +6,19 @@
 # Set Script Name variable
 SCRIPT=`basename ${BASH_SOURCE[0]}`
 
-# Find unused port
-PORT="1024"
-while $(lsof -Pi :$PORT -sTCP:LISTEN -t >/dev/null) ; do
-    PORT=$((PORT+1))
-done
+if [ -z "$PORT"]; then
+  # Find unused port
+  PORT="1024"
+  while $(lsof -Pi :$PORT -sTCP:LISTEN -t >/dev/null) ; do
+      PORT=$((PORT+1))
+  done
+fi
 
+SKYCOIN_NODE = "${SKYCOIN_NODE:~"127.0.0.1"}"
 COIN="${COIN:-skycoin}"
 RPC_PORT="$PORT"
-HOST="http://127.0.0.1:$PORT"
-RPC_ADDR="http://127.0.0.1:$RPC_PORT"
+HOST="http://$SKYCOIN_NODE:$PORT"
+RPC_ADDR="http://$SKYCOIN_NODE:$RPC_PORT"
 MODE="disable-wallet-api"
 BINARY="${COIN}-integration-disable-wallet-api.test"
 UPDATE=""
@@ -65,44 +68,46 @@ go test ./src/api/integration/...
 go test ./src/cli/integration/...
 
 DATA_DIR=$(mktemp -d -t ${COIN}-data-dir.XXXXXX)
-WALLET_DIR="${DATA_DIR}/wallets"
+WALLET_DIR="/wallets"
 
 if [[ ! "$DATA_DIR" ]]; then
   echo "Could not create temp dir"
   exit 1
 fi
 
-# Compile the skycoin node
-# We can't use "go run" because that creates two processes which doesn't allow us to kill it at the end
-echo "compiling $COIN with coverage"
-go test -c -ldflags "${GOLDFLAGS}" -tags testrunmain -o "$BINARY" -coverpkg="${COVERPKG}/..." ./cmd/${COIN}/
+if [ "$SKYCOIN_NODE" == "127.0.0.1" ]; then
+  # Compile the skycoin node
+  # We can't use "go run" because that creates two processes which doesn't allow us to kill it at the end
+  echo "compiling $COIN with coverage"
+  go test -c -ldflags "${GOLDFLAGS}" -tags testrunmain -o "$BINARY" -coverpkg="${COVERPKG}/..." ./cmd/${COIN}/
 
-mkdir -p coverage/
+  mkdir -p coverage/
 
-# Run skycoin node with pinned blockchain database
-echo "starting $COIN node in background with http listener on $HOST"
+  # Run skycoin node with pinned blockchain database
+  echo "starting $COIN node in background with http listener on $HOST"
 
-./"$BINARY" -disable-networking=true \
-            -web-interface-port=$PORT \
-            -download-peerlist=false \
-            -db-path=./src/api/integration/testdata/blockchain-180.db \
-            -db-read-only=true \
-            -launch-browser=false \
-            -data-dir="$DATA_DIR" \
-            -wallet-dir="$WALLET_DIR" \
-            -enable-all-api-sets=true \
-            -disable-api-sets=WALLET \
-            -test.run "^TestRunMain$" \
-            -test.coverprofile="${COVERAGEFILE}" \
-            &
+  ./"$BINARY" -disable-networking=true \
+              -web-interface-port=$PORT \
+              -download-peerlist=false \
+              -db-path=./src/api/integration/testdata/blockchain-180.db \
+              -db-read-only=true \
+              -launch-browser=false \
+              -data-dir="$DATA_DIR" \
+              -wallet-dir="$WALLET_DIR" \
+              -enable-all-api-sets=true \
+              -disable-api-sets=WALLET \
+              -test.run "^TestRunMain$" \
+              -test.coverprofile="${COVERAGEFILE}" \
+              &
 
-SKYCOIN_PID=$!
+  SKYCOIN_PID=$!
 
-echo "$COIN node pid=$SKYCOIN_PID"
+  echo "$COIN node pid=$SKYCOIN_PID"
 
-echo "sleeping for startup"
-sleep 3
-echo "done sleeping"
+  echo "sleeping for startup"
+  sleep 3
+  echo "done sleeping"
+fi
 
 set +e
 
@@ -111,12 +116,14 @@ SKYCOIN_INTEGRATION_TESTS=1 SKYCOIN_INTEGRATION_TEST_MODE=$MODE SKYCOIN_NODE_HOS
 
 FAIL=$?
 
-echo "shutting down $COIN node"
+if [ "$SKYCOIN_NODE" == "127.0.0.1" ]; then
+  echo "shutting down $COIN node"
 
-# Shutdown skycoin node
-kill -s SIGINT $SKYCOIN_PID
-wait $SKYCOIN_PID
+  # Shutdown skycoin node
+  kill -s SIGINT $SKYCOIN_PID
+  wait $SKYCOIN_PID
 
-rm "$BINARY"
+  rm "$BINARY"
+fi
 
 exit $FAIL
